Restructure code under src package
